flask-admin-api/requirements.txt
aniso8601==8.1.0
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-JSONRPC==1.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
pkg-resources==0.0.0
pytz==2020.5
requests==2.25.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
typeguard==2.10.0
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
